+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 All notable changes to the "hexdiff" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## Version 1.0.231115
+- Add keywords
+- Add changelog
+- Show hexdiff status bar on hexdiff document
+- User setting(bytes/line, 4byte print)
 
+## Version 1.0.231114
 - Initial release